feat(cdp): improve folder structure
