+++ plugin-server/src/cdp/metrics/metrics.ts
@@ -1,0 +1,37 @@
+import { Counter, Gauge, Histogram } from 'prom-client'
+
+export const histogramKafkaBatchSize = new Histogram({
+    name: 'cdp_function_executor_batch_size',
+    help: 'The size of the batches we are receiving from Kafka',
+    buckets: [0, 50, 100, 250, 500, 750, 1000, 1500, 2000, 3000, Infinity],
+})
+
+export const histogramKafkaBatchSizeKb = new Histogram({
+    name: 'cdp_function_executor_batch_size_kb',
+    help: 'The size in kb of the batches we are receiving from Kafka',
+    buckets: [0, 128, 512, 1024, 5120, 10240, 20480, 51200, 102400, 204800, Infinity],
+})
+
+export const counterFunctionInvocation = new Counter({
+    name: 'cdp_function_invocation',
+    help: 'A function invocation was evaluated with an outcome',
+    labelNames: ['outcome'], // One of 'failed', 'succeeded', 'overflowed', 'disabled', 'filtered'
+})
+
+export const counterParseError = new Counter({
+    name: 'cdp_function_parse_error',
+    help: 'A function invocation was parsed with an error',
+    labelNames: ['error'],
+})
+
+export const gaugeBatchUtilization = new Gauge({
+    name: 'cdp_cyclotron_batch_utilization',
+    help: 'Indicates how big batches are we are processing compared to the max batch size. Useful as a scaling metric',
+    labelNames: ['queue'],
+})
+
+export const counterJobsProcessed = new Counter({
+    name: 'cdp_cyclotron_jobs_processed',
+    help: 'The number of jobs we are managing to process',
+    labelNames: ['queue'],
+})
